\name{areaRegionExample}
\docType{data}
\alias{areaRegionExample}
\title{A region with two holes}
\description{
Three 2-column matrices.  The first is a set of vertices of a boundary 
polygon encompassing a region.  The other two, hole1 and hole2, are holes
in the region.  This example dataset is used to illustrate the situation
where there are holes in the region where a density or nonparametric
regression is to be applied.  It is used in the example of the function
\code{areaRegion}.
}
\usage{
areaRegionExample
}
%
\format{Three 2-column numerical matrices}
%
\keyword{datasets}


