\name{polygon1}
\docType{data}
\alias{polygon1}
\title{Example boundary with causeway}
\description{
A 2x19 matrix of vertices for the boundary of a region representing a
lake almost divided in half by a causeway.  This was used in a simulation
in the paper of Barry and McIntyre (2011).
}
\usage{
polygon1
}
%
\format{A 2x19 numerical matrix}
%
\source{
Ronald P. Barry, Julie McIntyre.  Estimating animal densities and home
range in regions with irregular boundaries and holes:  A lattice-based
alternative to the kernel density estimator.
Ecological Modelling 222 (2011)  1666-1672.
}
\keyword{datasets}


