% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeHole.R
\name{removeHole}
\alias{removeHole}
\title{Removes holes from the region prior to density estimation.}
\usage{
removeHole(hole_poly, nodeFillingOutput)
}
\arguments{
\item{hole_poly}{A numerical matrix of vertices of the hole polygon.}

\item{nodeFillingOutput}{An object of type nodeFillingOutput, returned 
by nodeFilling or removeHole.}
}
\value{
An object of type nodeFillingOutput.
}
\description{
If a hole in a region is specified as a polygon, the function 
removeHole removes all nodes in the nodeFillingOutput that 
are contained in the hole. This function is called by 
nodeFilling, so it is generally not needed by users.
}
\author{
Ronald P. Barry
}
