\name{nparExample}
\docType{data}
\alias{nparExample}
\title{Example boundary with a grid of locations}
\description{
The first item, \code{polygon2}, is 35x2 matrix describing the boundary of 
a region.  The second, \code{grid2}, is a set of 59 locations for simulated
values of a response variable.  The third item, \code{Z}, is a vector of
responses.  This dataset was created to test and
illustrate the non-parametric lattice based regression estimator.  See the
example for function \code{createNparReg}.
}
\usage{
nparExample
}

\format{
Two matrices and a vector.  One matrix is 35x2, the other is 59x2.
}

\references{
Julie McIntyre, Ronald P. Barry (2018)  A Lattice-Based 
Smoother for Regions with Irregular Boundaries and Holes.  
Journal of Computational and Graphical Statistics.  In Press.
}
%
%
\keyword{datasets}


